ecn-svc-legacy/requirements.txt
argcomplete==1.9.3
base58==1.0.0
boto3==1.9.0
botocore==1.12.0
certifi==2018.8.24
cfn-flip==1.0.3
chardet==3.0.4
click==6.7
docutils==0.14
durationpy==0.5
Flask==1.0.2
Flask-Cors==3.0.6
future==0.16.0
hjson==3.0.1
idna==2.7
itsdangerous==0.24
Jinja2==2.10.1
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
MarkupSafe==1.0
mongoengine==0.15.3
numpy==1.15.4
Pillow==6.2.0
placebo==0.8.2
pymongo==3.7.1
python-dateutil==2.6.1
python-dotenv==0.10.3
python-slugify==1.2.4
PyYAML==5.1
requests==2.20.0
s3transfer==0.1.13
six==1.11.0
toml==0.9.6
tqdm==4.19.1
troposphere==2.3.3
Unidecode==1.0.22
urllib3==1.24.2
Werkzeug==0.15.3
wsgi-request-logger==0.4.6
zappa==0.47.0
